feat: Combining reducers and initial state
