+++ src/index.js
@@ -2,14 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom'
 import App from './components/app'
 import { BrowserRouter } from 'react-router-dom';
-import { createStore } from 'redux';
-import reducer from './reducers';
 import { Provider } from 'react-redux';
+import store from './store';
 
-const store = createStore(
-  reducer,
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-);
+
 
 ReactDOM.render(<BrowserRouter>
                     <Provider store={store}>
