+++ src/reducers/comments.js
@@ -1,0 +1,5 @@
+function comments (state = {}, action) {
+    return state;
+}
+
+export default comments;
