+++ src/reducers/posts.js
@@ -1,0 +1,5 @@
+function posts (state = {}, action) {
+    return state;
+}
+
+export default posts;
